+++ lib-multisrc/initmanga/src/eu/kanade/tachiyomi/multisrc/initmanga/InitManga.kt
@@ -1,7 +1,6 @@
 package eu.kanade.tachiyomi.multisrc.initmanga
 
 import android.util.Base64
-import eu.kanade.tachiyomi.multisrc.initmanga.aes.AesDecrypt
 import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.source.model.Filter
 import eu.kanade.tachiyomi.source.model.FilterList
